+++ lib/src/widgets/button/mx_button_style.dart
@@ -40,6 +40,43 @@ class MXButtonStyle {
         backgroundColor = _getColorByWarn(context, statusEnum);
         textColor = MXTheme.of(context).whiteColor;
         break;
+      case MXButtonThemeEnum.info:
+        backgroundColor = _getColorByInfo(context, statusEnum);
+        textColor = MXTheme.of(context).fontUsePrimaryColor;
+        break;
+    }
+
+    /// 将border的颜色设置为背景色
+    borderColor = backgroundColor;
+  }
+
+  /// 通过按钮的主题风格来获取对应的按钮状态样式
+  MXButtonStyle.getPlainStyleByTheme(
+    BuildContext context,
+    MXButtonThemeEnum theme,
+    MXButtonStatusEnum statusEnum,
+  ) {
+    switch (theme) {
+      case MXButtonThemeEnum.primary:
+        backgroundColor = _getColorByPlainPrimary(context, statusEnum);
+        textColor = MXTheme.of(context).brandPrimaryColor;
+        break;
+      case MXButtonThemeEnum.error:
+        backgroundColor = _getColorByPlainError(context, statusEnum);
+        textColor = MXTheme.of(context).errorPrimaryColor;
+        break;
+      case MXButtonThemeEnum.success:
+        backgroundColor = _getColorByPlainSuccess(context, statusEnum);
+        textColor = MXTheme.of(context).successPrimaryColor;
+        break;
+      case MXButtonThemeEnum.warn:
+        backgroundColor = _getColorByPlainWarn(context, statusEnum);
+        textColor = MXTheme.of(context).warnPrimaryColor;
+        break;
+      case MXButtonThemeEnum.info:
+        backgroundColor = _getColorByInfo(context, statusEnum);
+        textColor = MXTheme.of(context).fontUsePrimaryColor;
+        break;
     }
 
     /// 将border的颜色设置为背景色
@@ -65,6 +102,9 @@ class MXButtonStyle {
         break;
       case MXButtonThemeEnum.warn:
         textColor = _getColorByWarn(context, statusEnum);
+        break;
+      case MXButtonThemeEnum.info:
+        textColor = MXTheme.of(context).fontUsePrimaryColor;
         break;
     }
 
@@ -99,6 +139,9 @@ class MXButtonStyle {
       case MXButtonThemeEnum.warn:
         textColor = _getColorByWarn(context, statusEnum);
         break;
+      case MXButtonThemeEnum.info:
+        textColor = MXTheme.of(context).fontUsePrimaryColor;
+        break;
     }
 
     /// 统一设置外框按钮的背景色
@@ -123,6 +166,19 @@ class MXButtonStyle {
     }
   }
 
+  /// 根据按钮的状态来获取对应的品牌色
+  Color _getColorByPlainPrimary(
+      BuildContext context, MXButtonStatusEnum statusEnum) {
+    switch (statusEnum) {
+      case MXButtonStatusEnum.normal:
+        return MXTheme.of(context).brandColor1;
+      case MXButtonStatusEnum.disabled:
+        return MXTheme.of(context).brandColor1;
+      case MXButtonStatusEnum.click:
+        return MXTheme.of(context).brandColor2;
+    }
+  }
+
   /// 根据按钮的状态来获取对应的错误色
   Color _getColorByError(BuildContext context, MXButtonStatusEnum statusEnum) {
     switch (statusEnum) {
@@ -132,6 +188,19 @@ class MXButtonStyle {
         return MXTheme.of(context).errorDisabledColor;
       case MXButtonStatusEnum.click:
         return MXTheme.of(context).errorClickColor;
+    }
+  }
+
+  /// 根据按钮的状态来获取对应的错误色
+  Color _getColorByPlainError(
+      BuildContext context, MXButtonStatusEnum statusEnum) {
+    switch (statusEnum) {
+      case MXButtonStatusEnum.normal:
+        return MXTheme.of(context).errorColor1;
+      case MXButtonStatusEnum.disabled:
+        return MXTheme.of(context).errorColor1;
+      case MXButtonStatusEnum.click:
+        return MXTheme.of(context).errorColor2;
     }
   }
 
@@ -147,6 +216,31 @@ class MXButtonStyle {
     }
   }
 
+  /// 根据按钮的状态来获取对应的警告色
+  Color _getColorByPlainWarn(
+      BuildContext context, MXButtonStatusEnum statusEnum) {
+    switch (statusEnum) {
+      case MXButtonStatusEnum.normal:
+        return MXTheme.of(context).warnColor1;
+      case MXButtonStatusEnum.disabled:
+        return MXTheme.of(context).warnColor1;
+      case MXButtonStatusEnum.click:
+        return MXTheme.of(context).warnColor2;
+    }
+  }
+
+  /// 根据按钮的状态来获取对应的警告色
+  Color _getColorByInfo(BuildContext context, MXButtonStatusEnum statusEnum) {
+    switch (statusEnum) {
+      case MXButtonStatusEnum.normal:
+        return MXTheme.of(context).infoPrimaryColor;
+      case MXButtonStatusEnum.disabled:
+        return MXTheme.of(context).infoDisabledColor;
+      case MXButtonStatusEnum.click:
+        return MXTheme.of(context).infoColor3;
+    }
+  }
+
   /// 根据按钮的状态来获取对应的成功色
   Color _getColorBySuccess(
       BuildContext context, MXButtonStatusEnum statusEnum) {
@@ -157,6 +251,19 @@ class MXButtonStyle {
         return MXTheme.of(context).successDisabledColor;
       case MXButtonStatusEnum.click:
         return MXTheme.of(context).successClickColor;
+    }
+  }
+
+  /// 根据按钮的状态来获取对应的成功色
+  Color _getColorByPlainSuccess(
+      BuildContext context, MXButtonStatusEnum statusEnum) {
+    switch (statusEnum) {
+      case MXButtonStatusEnum.normal:
+        return MXTheme.of(context).successColor1;
+      case MXButtonStatusEnum.disabled:
+        return MXTheme.of(context).successColor1;
+      case MXButtonStatusEnum.click:
+        return MXTheme.of(context).successColor2;
     }
   }